requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
requests==2.32.3
python-dotenv==1.0.1
